Added routes and wired up Assignee screen
